example to new sda framework. NOTE add to remove example because of unexported fields / method in sda, wont duplicate methods...
